+++ NuguAgents/Sources/CapabilityAgents/MediaPlayer/MediaPlayerAgentSong.swift
@@ -64,7 +64,7 @@ public struct MediaPlayerAgentSong: Codable {
         title = try container.decodeIfPresent(String.self, forKey: .title)
         duration = try container.decodeIfPresent(String.self, forKey: .duration)
         issueDate = try container.decodeIfPresent(String.self, forKey: .issueDate)
-        etc = try container.decode([String: AnyHashable].self, forKey: .etc)
+        etc = try? container.decodeIfPresent([String: AnyHashable].self, forKey: .etc)
     }
     
     /// <#Description#>